Update for Python 3 support

+++ kegbot/pycore/flow_meter.py
@@ -1,23 +1,6 @@
-# Copyright 2012 Mike Wakerly <opensource@hoho.com>
-#
-# This file is part of the Pykeg package of the Kegbot project.
-# For more information on Pykeg or Kegbot, see http://kegbot.org/
-#
-# Pykeg is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 2 of the License, or
-# (at your option) any later version.
-#
-# Pykeg is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Pykeg.  If not, see <http://www.gnu.org/licenses/>.
-
 """Utility module for summing flows."""
 
+from builtins import object
 import logging
 
 class FlowMeter(object):
@@ -45,7 +28,7 @@ class FlowMeter(object):
     The value `ticks` is always saved as `_last_ticks` for use in the next
     report.
     """
-    ticks = long(ticks)
+    ticks = int(ticks)
     self._logger.info('SetTicks: ticks=%s last=%s total=%s' % (
         ticks, self._last_ticks, self._total_ticks))
